+++ src/main/java/de/maxhenkel/delivery/net/MessageShowTask.java
@@ -5,6 +5,7 @@ import de.maxhenkel.delivery.Main;
 import de.maxhenkel.delivery.gui.ContractContainer;
 import de.maxhenkel.delivery.gui.containerprovider.TaskContainerProvider;
 import de.maxhenkel.delivery.items.ModItems;
+import de.maxhenkel.delivery.tasks.Group;
 import de.maxhenkel.delivery.tasks.Task;
 import net.minecraft.network.PacketBuffer;
 import net.minecraft.util.text.TranslationTextComponent;
@@ -32,13 +33,18 @@ public class MessageShowTask implements Message<MessageShowTask> {
 
     @Override
     public void executeServerSide(NetworkEvent.Context context) {
-        Task task = Main.TASK_MANAGER.getTask(this.task);
+        try {
+            Group group = Main.getProgression(context.getSender()).getPlayerGroup(context.getSender().getUniqueID());
+            Task task = Main.TASK_MANAGER.getTask(this.task, group);
 
-        if (task == null) {
-            return;
+            if (task == null) {
+                return;
+            }
+
+            TaskContainerProvider.openGui(context.getSender(), task, new TranslationTextComponent(ModItems.CONTRACT.getTranslationKey()), ContractContainer::new);
+        } catch (Exception e) {
+
         }
-
-        TaskContainerProvider.openGui(context.getSender(), task, new TranslationTextComponent(ModItems.CONTRACT.getTranslationKey()), ContractContainer::new);
     }
 
     @Override